+++ urlfilter/decode_test.go
@@ -1,4 +1,4 @@
-package urlvalues_test
+package urlfilter_test
 
 import (
 	"database/sql"
@@ -8,7 +8,7 @@ import (
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
 
-	"github.com/go-pg/pg/v9/urlvalues"
+	"github.com/go-pg/pg/v9/urlfilter"
 )
 
 type CustomField struct {
@@ -45,14 +45,10 @@ type Filter struct {
 	Omit []byte `pg:"-"`
 }
 
-func (f *Filter) AfterDecodeURLValues(values urlvalues.Values) error {
-	return nil
-}
-
 var _ = Describe("Decode", func() {
 	It("decodes struct from Values", func() {
 		f := &Filter{}
-		err := urlvalues.Decode(f, urlvalues.Values{
+		err := urlfilter.Decode(f, urlfilter.Values{
 			"field":      {"one"},
 			"field__neq": {"two"},
 			"field__lt":  {"1"},
@@ -103,7 +99,7 @@ var _ = Describe("Decode", func() {
 
 	It("supports names with suffix `[]`", func() {
 		f := &Filter{}
-		err := urlvalues.Decode(f, urlvalues.Values{
+		err := urlfilter.Decode(f, urlfilter.Values{
 			"field[]": {"one"},
 		})
 		Expect(err).NotTo(HaveOccurred())
@@ -112,7 +108,7 @@ var _ = Describe("Decode", func() {
 
 	It("supports names with prefix `:`", func() {
 		f := &Filter{}
-		err := urlvalues.Decode(f, urlvalues.Values{
+		err := urlfilter.Decode(f, urlfilter.Values{
 			":field": {"one"},
 		})
 		Expect(err).NotTo(HaveOccurred())
@@ -121,7 +117,7 @@ var _ = Describe("Decode", func() {
 
 	It("decodes sql.Null*", func() {
 		f := &Filter{}
-		err := urlvalues.Decode(f, urlvalues.Values{
+		err := urlfilter.Decode(f, urlfilter.Values{
 			"null_bool":    {""},
 			"null_int64":   {""},
 			"null_float64": {""},